+++ lib/karafka/web/tracking/consumers/sampler.rb
@@ -14,7 +14,7 @@ module Karafka
           # Current schema version
           # This can be used in the future for detecting incompatible changes and writing
           # migrations
-          SCHEMA_VERSION = '1.0.2'
+          SCHEMA_VERSION = '1.1.0'
 
           # 60 seconds window for time tracked window-based metrics
           TIMES_TTL = 60